chore: rename 'distr' -> 'shell_completions'
