+++ crates/amaru/src/bin/amaru/cmd/shell_completions.rs
