+++ crates/amaru/src/bin/amaru/main.rs
@@ -59,7 +59,7 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
         cli::Command::Bootstrap(args) => cmd::bootstrap::run(args).await,
         cli::Command::FetchChainHeaders(args) => cmd::fetch_chain_headers::run(args).await,
         cli::Command::CreateSnapshots(args) => cmd::create_snapshots::run(args).await,
-        cli::Command::ShellCompletions(args) => cmd::distr::run(args).await,
+        cli::Command::ShellCompletions(args) => cmd::shell_completions::run(args).await,
         cli::Command::DumpChainDB(args) => cmd::dump_chain_db::run(args).await,
         cli::Command::RemoveValidationStatus(args) => cmd::remove_validation_status::run(args).await,
         cli::Command::RemoveChain(args) => cmd::remove_chain::run(args).await,
